modified PagedResult Count to be int64

+++ src/CavemanTools/Model/IPagedResult.cs
@@ -13,7 +13,7 @@ namespace CavemanTools.Model
 		/// <summary>
 		/// Gets the total number of items
 		/// </summary>
-		int Count { get;}
+		long Count { get;}
 
 		/// <summary>
 		/// Gets result items
